backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0

SQLAlchemy==2.0.36
alembic==1.14.0

psycopg2-binary>=2.9.10

pydantic>=2.10,<3
pydantic-settings>=2.7,<3
email-validator==2.2.0

python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0

